requirements.txt
certifi>=2024.7.4
charset-normalizer>=3.3.2
idna >=3.7
meraki>=1.49.0
PyYAML>=6.0.2
requests>=2.32.3
urllib3>=2.2.2
